refactor: migrate shared entries to TypeScript
