+++ lib/index.ts
@@ -1,5 +1,2 @@
-// Components.
 export * from './components';
-
-// Component-related.
 export { getColumnConfigOutputText } from './components/slider/utils';
