+++ lib/utilities/web-workers.ts
@@ -1,12 +1,26 @@
+import type { MutableRefObject } from 'react';
+
+interface WorkerMessage {
+	success: boolean;
+	data?: unknown;
+	error?: string;
+}
+
+interface AnalysisSettings {
+	[key: string]: unknown;
+}
+
 /**
  * Returns a memoized function that gets or creates a singleton image analysis worker (with inline fallback).
- * Also returns a callback for analyzing with the worker.
- * @param {object} workerRef - React ref object for the worker instance.
- * @param {string} workerInline - Inlined worker code string.
- * @returns {{ getOrCreateWorker: Function, analyzeWithWorkerCb: Function }}
  */
-export function useImageAnalysisWorker(workerRef, workerInline) {
-	const getOrCreateWorker = () => {
+export function useImageAnalysisWorker(
+	workerRef: MutableRefObject<Worker | null>,
+	workerInline: string,
+): {
+	getOrCreateWorker: () => Worker;
+	analyzeWithWorkerCb: (imageBitmap: ImageBitmap, settings: AnalysisSettings) => Promise<unknown>;
+} {
+	const getOrCreateWorker = (): Worker => {
 		if (!workerRef.current) {
 			workerRef.current = createImageAnalysisWorker(workerInline);
 			workerRef.current.addEventListener('error', (e) => {
@@ -17,7 +31,7 @@ export function useImageAnalysisWorker(workerRef, workerInline) {
 		return workerRef.current;
 	};
 
-	const analyzeWithWorkerCb = async (imageBitmap, settings) => {
+	const analyzeWithWorkerCb = async (imageBitmap: ImageBitmap, settings: AnalysisSettings): Promise<unknown> => {
 		const worker = getOrCreateWorker();
 
 		return analyzeWithWorker(worker, imageBitmap, settings);
@@ -26,15 +40,10 @@ export function useImageAnalysisWorker(workerRef, workerInline) {
 	return { getOrCreateWorker, analyzeWithWorkerCb };
 }
 
-// Utilities for working with web workers in the Eightshift UI Components library
-// Extracted from SmartImage for reuse and clarity
-
 /**
  * Create a new image analysis worker using the inlined worker code.
- * @param {string} workerInline - The inlined worker code as a string.
- * @returns {Worker}
  */
-export function createImageAnalysisWorker(workerInline) {
+export function createImageAnalysisWorker(workerInline: string): Worker {
 	if (!workerInline || typeof workerInline !== 'string' || workerInline.length < 100) {
 		throw new Error('Worker could not be created: inline worker code not available. Make sure the worker is properly bundled.');
 	}
@@ -46,14 +55,10 @@ export function createImageAnalysisWorker(workerInline) {
 
 /**
  * Analyze an image using a web worker (returns a Promise).
- * @param {Worker} worker - The worker instance.
- * @param {ImageBitmap} imageBitmap - The image to analyze.
- * @param {object} settings - Analysis settings.
- * @returns {Promise<object>} - Resolves with analysis result.
  */
-export function analyzeWithWorker(worker, imageBitmap, settings) {
+export function analyzeWithWorker(worker: Worker, imageBitmap: ImageBitmap, settings: AnalysisSettings): Promise<unknown> {
 	return new Promise((resolve, reject) => {
-		const handleMessage = (event) => {
+		const handleMessage = (event: MessageEvent<WorkerMessage>) => {
 			const { success, data, error } = event.data;
 
 			if (success) {
@@ -65,11 +70,13 @@ export function analyzeWithWorker(worker, imageBitmap, settings) {
 			worker.removeEventListener('message', handleMessage);
 			worker.removeEventListener('error', handleError);
 		};
-		const handleError = (error) => {
-			reject(error);
+
+		const handleError = (error: ErrorEvent) => {
+			reject(new Error(error.message || 'Image analysis worker failed.'));
 			worker.removeEventListener('message', handleMessage);
 			worker.removeEventListener('error', handleError);
 		};
+
 		worker.addEventListener('message', handleMessage);
 		worker.addEventListener('error', handleError);
 		worker.postMessage({ imageBitmap, settings }, [imageBitmap]);
